requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
pymysql==1.1.1
alembic==1.13.2
python-dotenv==1.0.1
pydantic==1.10.18
httpx==0.27.2
apscheduler==3.10.4
PyYAML==6.0.2
